fix: cleanup matrix data response
